requirements.txt
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
matplotlib==3.7.2